Refactor: Replace auth_errors with auth_exception in relevant files and implement AuthException hierarchy
