+++ lib/core/errors/auth_exception.dart
