+++ lib/core/utils/services/auth/firebase_auth_service.dart
@@ -1,5 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:skillbridge/core/errors/auth_errors.dart';
+import 'package:skillbridge/core/errors/auth_exception.dart';
 import 'package:skillbridge/core/models/auth_user_model.dart';
 import 'package:skillbridge/core/utils/services/auth/firebase_auth_service_repo.dart';
 import 'package:skillbridge/core/utils/validator/app_validator.dart';
